package/usb_modeswitch: fix installation of systemd unit file
The upstream Makefile failed to detect systemd being present in some (maybe all)
builds, resulting in the unit file not being installed. Without the unit file,
the udev rules in usb_modeswitch-data don't work as expected (no modeswitch is
performed).

This commit adds a patch that modifies the Makefile to include
'$(PREFIX)/bin/systemctl' in the list of paths checked, which makes the
installation phase work as intended. I will also submit this patch upstream.

Signed-off-by: Sol Bekic <s+removethis@s-ol.nu>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>