package/x11r7/xdriver_xf86-video-amdgpu: needs libgbm
The package depends on gbm:
https://gitlab.freedesktop.org/xorg/driver/xf86-video-amdgpu/-/blob/master/configure.ac#L75

This dependency was always there. However, previously it was implicit
from the mesa3d dependencies. Since we made libgbm a virtual package, it
is possible that libgbm is selected in another way. Note that the
original dependencies would always for libgbm in mesa3d, so this was not
causing any build failures.

Also explicitly add libgbm to _DEPENDENCIES. Previously, this was
implied by the mesa3d dependency in xserver, but this is no longer true
for other gbm providers.

For details see
http://lists.busybox.net/pipermail/buildroot/2022-April/641451.html

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>