package/network-manager: put back libuuid dependency
Commit 917446d262b3334476debc2c7a0ecf18a00ce114 wrongly removed libuuid
dependency resulting in the following build failure:

checking for UUID... no
configure: error: Package requirements (uuid) were not met:

Package 'uuid', required by 'virtual:world', not found

It should be noted that <uuid/uuid.h> is still used in one example:
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/blob/main/examples/C/glib/add-connection-gdbus.c#L17
as stated in
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/73b9883c6f30320b727a664ff2c84cc772aac94b:

"We still require libuuid for building, because it is used by an example
program. Maybe that should be changed, to avoid the build dependency.
But that can be done at a later time."

Fixes:
 - http://autobuild.buildroot.org/results/ac25c42d73161bf4cdfae3e53de8826f0b8dbf9c

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>