package/python-pillow: fix host package build
In commit
f38638706ac27d7babf3b1d6868486c66eb866ef ("package/python-pillow:
enable host variant"), when support for the host variant of
python-pillow was added, the fact that python-pillow has optional
dependencies was not handled, and left python-pillow to autodetect its
optional dependencies.

Unfortunately, without some config options, some of its dependencies
are made mandatory, such as the JPEG library. Since for now such
support is not needed in the host variant of python-pillow, we simply
disable all optional dependencies.

To be consistent with the target variant, we also ensure to have
host-pkgconf as a dependency of the host variant.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>