docs/manual: update documentation about partition-type-uuid convention in genimage.cfg
Whenever possible, we want to use the "shortcuts" provided by genimage
for the partition-type-uuid values:

 - U for the EFI System Partition
 - F for FAT32 partitions
 - L for root filesystem partitions and other Linux filesystems

In follow-up commits, we will convert existing genimage.cfg to this
convention, except for the root filesystem, if the existing
genimage.cfg specifies a partition-type-uuid that doesn't match the
one that "L" would imply. In this case, we will keep things as-is to
not break existing setups. However, for all future genimage.cfg files,
we recommend to use "L" for the root filesystem.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>