package/libkrb5: fix LICENSE list
The NOTICE file contains a long list of licenses. A few of them are MIT,
BSD-2-Clause, BSD-3-Clause, or BSD-4-Clause. And then there are many
that are somewhat similar but not exactly.

Mention all the "known" licenses in LIBKRB5_LICENSE, and add "others" to
indicate that there is more.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>