package/ntpsec: fix dependencies
Add python3 dependencies to avoid the following build failure raised
since the addition of the package in commit
d6f0992ba6e5526a8eae53870a755803d5e4ca50:

Makefile:576: *** libffi is in the dependency chain of python3 that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

Fixes:
 - http://autobuild.buildroot.org/results/2893c213f49d37a0878dcf14ec68ef991471c70f

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>