package/avocado: fix license
In commit a4c830a0ff58b636d13fca1aa339c0d6e331cb85 ("package/avocado:
new package") the license information that was introduced was
incorrect. The license is GPL-2.0, not Apache-2.0.

According to the LICENSE file:

Most files in Avocado are released under the GPLv2 with a "and any later
version" clause (GPLv2+), but some files are explicitly bound to the
version 2 of the license and therefore that's the official license of
the project itself. The header on each individual file contains
information about the GPL version that applies to it.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>