package/mesa3d-headers: bring back the installation of dri.pc
With the version bump to 22.x, the dri.pc file was no longer deployed as
it was believed to not be provided.

While it was temporarily missing from upstream as part of deprecating
the classic driver, it was reintroduced prior to the 22.0 release via:
  https://gitlab.freedesktop.org/mesa/mesa/-/commit/3ae3569d822c2546c4c748090f8a6c8016f763ea

This file is still necessary to build Xorg server. See previous commits:
  db92c6272e
  7468b60e7c

Fixes: 5ddc98c3f ("package/{mesa3d, mesa3d-headers}: bump version to 22.1.3")

Signed-off-by: Vincent Fazio <vfazio@xes-inc.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>